backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
timm
faiss-cpu
sentence-transformers
requests
